Bugfix: Fix syntax
`if: ${{ inputs.transform-configuration }} == true` is always true - this syntax is not correct for the intention. See Github Action's documentation for [an example of testing equality of a configuration value](https://docs.github.com/en/actions/learn-github-actions/expressions#example-for-workflow-step).

This PR also makes the `transform-configuration` setting optional, and defaults it to false for backwards compatibility.